+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This API sends notifications to users for things like forgotten passwords, initi
 ### Changed
 - YLP-559 emails which are part of URLs are URL encoded
 - YLP-516 invitations to join a team
+### Fixed
+- YLP-532 Reset password demand can be used several times
+
 ## 1.5.1 - 2021-03-12
 ### Engineering Use
 - Switch CI build to Jenkins (instead of Travis)